[change] 使用int64存储datetime类型，避免读取日期超过2038年出现问题

+++ src/Luban.Job.Common/Source/TypeVisitors/CsUnderingDefineTypeName.cs
@@ -152,7 +152,7 @@ namespace Luban.Job.Common.TypeVisitors
             {
                 return mapper.TargetTypeName;
             }
-            return "int";
+            return "long";
         }
     }
 }
